Remove slot tree item base
