+++ apps/vs-code-designer/src/app/commands/browseWebsite.ts
@@ -5,11 +5,10 @@
 import { logicAppFilter } from '../../constants';
 import { ext } from '../../extensionVariables';
 import type { SlotTreeItem } from '../tree/slotsTree/SlotTreeItem';
-import type { SlotTreeItemBase } from '../tree/slotsTree/SlotTreeItemBase';
 import type { IActionContext } from '@microsoft/vscode-azext-utils';
 import { openUrl } from '@microsoft/vscode-azext-utils';
 
-export async function browseWebsite(context: IActionContext, node?: SlotTreeItemBase): Promise<void> {
+export async function browseWebsite(context: IActionContext, node?: SlotTreeItem): Promise<void> {
   if (!node) {
     node = await ext.rgApi.pickAppResource<SlotTreeItem>(context, {
       filter: logicAppFilter,
