+++ apps/vs-code-designer/src/app/commands/configureDeploymentSource.ts
@@ -5,11 +5,10 @@
 import { logicAppFilter } from '../../constants';
 import { ext } from '../../extensionVariables';
 import type { SlotTreeItem } from '../tree/slotsTree/SlotTreeItem';
-import type { SlotTreeItemBase } from '../tree/slotsTree/SlotTreeItemBase';
 import { editScmType } from '@microsoft/vscode-azext-azureappservice';
 import type { IActionContext } from '@microsoft/vscode-azext-utils';
 
-export async function configureDeploymentSource(context: IActionContext, node?: SlotTreeItemBase): Promise<void> {
+export async function configureDeploymentSource(context: IActionContext, node?: SlotTreeItem): Promise<void> {
   if (!node) {
     node = await ext.rgApi.pickAppResource<SlotTreeItem>(context, {
       filter: logicAppFilter,
