+++ apps/vs-code-designer/src/app/commands/deploy/updateAppSettings.ts
@@ -1,5 +1,5 @@
 import { workflowAppAADClientId, workflowAppAADClientSecret, workflowAppAADObjectId, workflowAppAADTenantId } from '../../../constants';
-import type { SlotTreeItemBase } from '../../tree/slotsTree/SlotTreeItemBase';
+import type { SlotTreeItem } from '../../tree/slotsTree/SlotTreeItem';
 import type { StringDictionary } from '@azure/arm-appservice';
 import type { IActionContext } from '@microsoft/vscode-azext-utils';
 import type { IIdentityWizardContext } from '@microsoft/vscode-extension';
@@ -7,12 +7,12 @@ import type { IIdentityWizardContext } from '@microsoft/vscode-extension';
 /**
  * Updates remote logic app settings with identity details.
  * @param {IActionContext} context - Command context.
- * @param {SlotTreeItemBase} node - Logic app node structure.
+ * @param {SlotTreeItem} node - Logic app node structure.
  * @param {IIdentityWizardContext} identityWizardContext - Identity context.
  */
 export async function updateAppSettingsWithIdentityDetails(
   context: IActionContext,
-  node: SlotTreeItemBase,
+  node: SlotTreeItem,
   identityWizardContext: IIdentityWizardContext
 ): Promise<void> {
   const client = await node.site.createClient(context);
