+++ apps/vs-code-designer/src/app/commands/remoteDebug/startRemoteDebug.ts
@@ -4,16 +4,16 @@
  *--------------------------------------------------------------------------------------------*/
 import { logicAppFilter } from '../../../constants';
 import { ext } from '../../../extensionVariables';
-import type { SlotTreeItemBase } from '../../tree/slotsTree/SlotTreeItemBase';
+import type { SlotTreeItem } from '../../tree/slotsTree/SlotTreeItem';
 import { getRemoteDebugLanguage } from './getRemoteDebugLanguage';
 import type { SiteConfig } from '@azure/arm-appservice';
 import * as appservice from '@microsoft/vscode-azext-azureappservice';
 import type { IActionContext } from '@microsoft/vscode-azext-utils';
 import * as vscode from 'vscode';
 
-export async function startRemoteDebug(context: IActionContext, node?: SlotTreeItemBase): Promise<void> {
+export async function startRemoteDebug(context: IActionContext, node?: SlotTreeItem): Promise<void> {
   if (!node) {
-    node = await ext.rgApi.pickAppResource<SlotTreeItemBase>(context, {
+    node = await ext.rgApi.pickAppResource<SlotTreeItem>(context, {
       filter: logicAppFilter,
     });
   }
