+++ apps/vs-code-designer/src/app/commands/restartLogicApp.ts
@@ -2,12 +2,12 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-import type { SlotTreeItemBase } from '../tree/slotsTree/SlotTreeItemBase';
+import type { SlotTreeItem } from '../tree/slotsTree/SlotTreeItem';
 import { startLogicApp } from './startLogicApp';
 import { stopLogicApp } from './stopLogicApp';
 import type { IActionContext } from '@microsoft/vscode-azext-utils';
 
-export async function restartLogicApp(context: IActionContext, node?: SlotTreeItemBase): Promise<void> {
+export async function restartLogicApp(context: IActionContext, node?: SlotTreeItem): Promise<void> {
   node = await stopLogicApp(context, node);
   await startLogicApp(context, node);
 }
